some functions moved from MobileDormantController.kt and FixedController.kt to SessionsController.kt
