+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/active/MobileActiveController.kt
@@ -32,7 +32,7 @@ class MobileActiveController(
     mSettings: Settings,
     mApiServiceFactory: ApiServiceFactory,
     private val airBeamReconnector: AirBeamReconnector
-): SessionsController(mRootActivity, mViewMvc, mSessionsViewModel, mSettings, mApiServiceFactory),
+): SessionsController(mRootActivity, mViewMvc, mSessionsViewModel, mSettings, mApiServiceFactory, mRootActivity!!.supportFragmentManager),
     SessionsViewMvc.Listener {
 
     private var mSessionsObserver = ActiveSessionsObserver(mLifecycleOwner, mSessionsViewModel, mViewMvc)
@@ -84,4 +84,13 @@ class MobileActiveController(
             }
         })
     }
+
+    override fun onEditDataPressed() { // Edit session bottom sheet handling
+        // do nothing
+    }
+
+    override fun onCancelPressed() {
+        // do nothing
+    }
+
 }
