+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/dormant/MobileDormantController.kt
@@ -28,12 +28,12 @@ class MobileDormantController(
     mLifecycleOwner: LifecycleOwner,
     mSettings: Settings,
     mApiServiceFactory: ApiServiceFactory,
-    private val fragmentManager: FragmentManager
-): SessionsController(mRootActivity, mViewMvc, mSessionsViewModel, mSettings, mApiServiceFactory),
+    fragmentManager: FragmentManager
+): SessionsController(mRootActivity, mViewMvc, mSessionsViewModel, mSettings, mApiServiceFactory, fragmentManager),
     SessionsViewMvc.Listener, EditSessionBottomSheet.Listener {
 
     private var mSessionsObserver = DormantSessionsObserver(mLifecycleOwner, mSessionsViewModel, mViewMvc)
-    private var dialog: EditSessionBottomSheet? = null
+
 
     override fun registerSessionsObserver() {
         mSessionsObserver.observe(mSessionsViewModel.loadMobileDormantSessionsWithMeasurements())
@@ -47,10 +47,6 @@ class MobileDormantController(
         startNewSession(Session.Type.MOBILE)
     }
 
-    override fun onEditSessionClicked(session: Session) {
-        startEditSessionBottomSheet(session)
-    }
-
     override fun onDeleteSessionClicked(sessionUUID: String) {
         val event = DeleteSessionEvent(sessionUUID)
         EventBus.getDefault().post(event)
@@ -59,30 +55,5 @@ class MobileDormantController(
     override fun onStopSessionClicked(sessionUUID: String) {
         // do nothing
     }
-
-    private fun startEditSessionBottomSheet(session: Session) {
-        dialog = EditSessionBottomSheet(this, session)
-        dialog?.show(fragmentManager, "MobileDormantEdit")
-    }
-
-    override fun onEditDataPressed() { // handling buttons in EditSessionBottomSheet
-        val editData = dialog?.editDataConfirmed()
-        if(editData == null){
-            Log.e("EDIT_SESS", "Edit data is null")
-        }else{
-            editSessionEventPost(editData)
-        }
-        dialog?.dismiss()
-    }
-
-    override fun onCancelPressed() { // handling buttons in EditSessionBottomSheet
-        dialog?.dismiss()
-    }
-
-    fun editSessionEventPost(session: Session){
-        val event = EditSessionEvent(session)
-        EventBus.getDefault().post(event)
-    }
-
 
 }
